+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/send/SendFragment.kt
@@ -142,7 +142,6 @@ class SendFragment : BaseFragment() {
                                 findNavController(),
                                 sendEvmViewModel,
                                 amountInputModeViewModel,
-                                sendEntryPointDestId,
                                 prefilledData,
                             )
                         }
