+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/sendevmtransaction/SendEvmTransactionView.kt
@@ -1,5 +1,8 @@
 package io.horizontalsystems.bankwallet.modules.sendevmtransaction
 
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -12,6 +15,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalView
@@ -21,12 +25,18 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.ethereum.CautionViewItem
 import io.horizontalsystems.bankwallet.core.iconPlaceholder
 import io.horizontalsystems.bankwallet.core.imageUrl
 import io.horizontalsystems.bankwallet.core.shorten
+import io.horizontalsystems.bankwallet.core.slideFromBottom
 import io.horizontalsystems.bankwallet.modules.evmfee.Cautions
 import io.horizontalsystems.bankwallet.modules.evmfee.EvmFeeCellViewModel
+import io.horizontalsystems.bankwallet.modules.evmfee.FeeSettingsInfoDialog
 import io.horizontalsystems.bankwallet.modules.fee.FeeCell
+import io.horizontalsystems.bankwallet.modules.multiswap.QuoteInfoRow
+import io.horizontalsystems.bankwallet.modules.multiswap.ui.DataField
+import io.horizontalsystems.bankwallet.modules.send.SendModule
 import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmNonceViewModel
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.ButtonSecondaryDefault
@@ -40,6 +50,7 @@ import io.horizontalsystems.bankwallet.ui.compose.components.TransactionInfoAddr
 import io.horizontalsystems.bankwallet.ui.compose.components.TransactionInfoContactCell
 import io.horizontalsystems.bankwallet.ui.compose.components.VSpacer
 import io.horizontalsystems.bankwallet.ui.compose.components.caption_grey
+import io.horizontalsystems.bankwallet.ui.compose.components.cell.SectionUniversalLawrence
 import io.horizontalsystems.bankwallet.ui.compose.components.headline2_leah
 import io.horizontalsystems.bankwallet.ui.compose.components.subhead1_grey
 import io.horizontalsystems.bankwallet.ui.compose.components.subhead1_leah
@@ -54,14 +65,82 @@ import io.horizontalsystems.marketkit.models.Token
 import io.horizontalsystems.marketkit.models.TokenType
 
 @Composable
+fun SendEvmTransactionViewNew(
+    navController: NavController,
+    items: List<SectionViewItem>,
+    cautions: List<CautionViewItem>,
+    transactionFields: List<DataField>,
+    networkFee: SendModule.AmountData?,
+) {
+    Column {
+        items.forEach { sectionViewItem ->
+            SectionView(sectionViewItem.viewItems, navController)
+        }
+
+        if (transactionFields.isNotEmpty()) {
+            VSpacer(height = 16.dp)
+            SectionUniversalLawrence {
+                transactionFields.forEachIndexed { index, field ->
+                    field.GetContent(navController, index != 0)
+                }
+            }
+        }
+
+        VSpacer(height = 16.dp)
+        SectionUniversalLawrence {
+            QuoteInfoRow(
+                title = {
+                    val title = stringResource(id = R.string.FeeSettings_NetworkFee)
+                    val infoText = stringResource(id = R.string.FeeSettings_NetworkFee_Info)
+
+                    subhead2_grey(text = title)
+
+                    Image(
+                        modifier = Modifier
+                            .padding(horizontal = 8.dp)
+                            .clickable(
+                                onClick = {
+                                    navController.slideFromBottom(
+                                        R.id.feeSettingsInfoDialog,
+                                        FeeSettingsInfoDialog.Input(title, infoText)
+                                    )
+                                },
+                                interactionSource = remember { MutableInteractionSource() },
+                                indication = null
+                            )
+                        ,
+                        painter = painterResource(id = R.drawable.ic_info_20),
+                        contentDescription = ""
+                    )
+
+                },
+                value = {
+                    val primary = networkFee?.primary?.getFormattedPlain() ?: "---"
+                    val secondary = networkFee?.secondary?.getFormattedPlain() ?: "---"
+
+                    Column(horizontalAlignment = Alignment.End) {
+                        subhead2_leah(text = primary)
+                        VSpacer(height = 1.dp)
+                        subhead2_grey(text = secondary)
+                    }
+                }
+            )
+        }
+
+        if (cautions.isNotEmpty()) {
+            Cautions(cautions)
+        }
+    }
+}
+
+@Composable
 fun SendEvmTransactionView(
-    transactionViewModel: SendEvmTransactionViewModel,
     feeCellViewModel: EvmFeeCellViewModel,
     nonceViewModel: SendEvmNonceViewModel,
     navController: NavController,
+    items: List<SectionViewItem>,
+    cautions: List<CautionViewItem>?,
 ) {
-
-    val items by transactionViewModel.viewItemsLiveData.observeAsState(listOf())
     val fee by feeCellViewModel.feeLiveData.observeAsState(null)
     val viewState by feeCellViewModel.viewStateLiveData.observeAsState()
 
@@ -85,7 +164,6 @@ fun SendEvmTransactionView(
             }
         )
 
-        val cautions by transactionViewModel.cautionsLiveData.observeAsState()
         cautions?.let {
             Cautions(it)
         }
